Create README.md
This application system is a Unizulu infrastructure reporting and maintenance system for students and staff members of the University of Zululand. This software  application allows students and staff members to report any faulty property that is on-campus. The system is designed to reduce the number of broken infrastructure by bridging the gap of communication between staff, students and warden that will enable the PPO to fix broken infrastructure immediately after the report being sent by student or a staff member without any further harm.
The software will be solely based communication between students, staff members, and warden via online. Reporting with smartphones in the application. More specifically the system is designed to allow students and staff to report any faulty infrastructure to warden thus the warden will call the PPO help desk to inform them about a damage infrastructure on-campus. This system will also have a database containing a list of University of Zululand staff members and students that have registered to use this application.